Add an OSS build flavor.
The OSS build flavor doesn't include proprietary dependencies such as
Firebase or GMS, which is better privacy-wise. This opens the
possibility to submitting the app to the F-Droid app store.

Unfortunately, proprietary Firebase and GMS gradle plugins are
**still** required to build the OSS flavor. There doesn't seem to be
an easy way around this when using Kotlin DSL as those plugins declare
functions that must be in classpath to build the gradle scripts. We'll
strip those out in the F-Droid receipe if required.